requirements.txt
tweepy>=4.10.0
natsort>=8.1.0
Pillow>=9.2.0
moviepy>=1.0.3